+++ systemtest/src/main/java/io/strimzi/systemtest/utils/kafkaUtils/KafkaConnectorUtils.java
@@ -4,10 +4,14 @@
  */
 package io.strimzi.systemtest.utils.kafkaUtils;
 
+import io.strimzi.api.kafka.Crds;
 import io.strimzi.systemtest.Constants;
 import io.strimzi.test.TestUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+
+import static io.strimzi.test.k8s.KubeClusterResource.cmdKubeClient;
+import static io.strimzi.test.k8s.KubeClusterResource.kubeClient;
 
 public class KafkaConnectorUtils {
 
@@ -27,7 +31,7 @@ public class KafkaConnectorUtils {
 
         TestUtils.waitFor("Waiting for stability of connector " + connectorName, Constants.GLOBAL_POLL_INTERVAL, Constants.GLOBAL_TIMEOUT,
             () -> {
-                String availableConnectors = KafkaConnectUtils.getCreatedConnectors(connectPodName);
+                String availableConnectors = getCreatedConnectors(connectPodName);
                 if (availableConnectors.contains(connectorName)) {
                     LOGGER.info("Connector with name {} is present. Remaining seconds for stability {}", connectorName,
                             Constants.GLOBAL_RECONCILIATION_COUNT - i[0]);
@@ -37,5 +41,26 @@ public class KafkaConnectorUtils {
                 }
             }
         );
+    }
+
+    public static void waitForConnectorStatus(String name, String state) {
+        LOGGER.info("Waiting for Kafka Connector {}", name);
+        TestUtils.waitFor(" Kafka Connector " + name + " is ready", Constants.POLL_INTERVAL_FOR_RESOURCE_READINESS, Constants.TIMEOUT_FOR_RESOURCE_READINESS,
+                () -> Crds.kafkaConnectorOperation(kubeClient().getClient()).inNamespace(kubeClient().getNamespace()).withName(name).get().getStatus().getConditions().get(0).getType().equals(state));
+        LOGGER.info("Kafka Connector {} is ready", name);
+    }
+
+
+    public static String getCreatedConnectors(String connectPodName) {
+        return cmdKubeClient().execInPod(connectPodName, "/bin/bash", "-c",
+                "curl -X GET http://localhost:8083/connectors"
+        ).out();
+    }
+
+    public static void waitForConnectorCreation(String connectS2IPodName, String connectorName) {
+        TestUtils.waitFor(connectorName + " connector creation", Constants.GLOBAL_POLL_INTERVAL, Constants.GLOBAL_STATUS_TIMEOUT, () -> {
+            String availableConnectors = getCreatedConnectors(connectS2IPodName);
+            return availableConnectors.contains(connectorName);
+        });
     }
 }
